Rename substitution config admin plugin to interpolation.
git-svn-id: https://svn.apache.org/repos/asf/felix/trunk@1863793 13f79535-47bb-0310-9956-ffa450edef68
